+++ src/main/java/org/jabref/logic/exporter/SavePreferences.java
@@ -3,7 +3,7 @@ package org.jabref.logic.exporter;
 import java.nio.charset.Charset;
 
 import org.jabref.logic.bibtex.FieldWriterPreferences;
-import org.jabref.logic.bibtexkeypattern.CitationKeyPatternPreferences;
+import org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences;
 import org.jabref.model.metadata.SaveOrderConfig;
 
 public class SavePreferences {
