+++ src/main/java/org/jabref/logic/integrity/ValidCitationKeyChecker.java
@@ -2,7 +2,7 @@ package org.jabref.logic.integrity;
 
 import java.util.Optional;
 
-import org.jabref.logic.bibtexkeypattern.CitationKeyGenerator;
+import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.strings.StringUtil;
