+++ src/test/java/org/jabref/logic/citationkeypattern/BracketedPatternTest.java
@@ -1,4 +1,4 @@
-package org.jabref.logic.bibtexkeypattern;
+package org.jabref.logic.citationkeypattern;
 
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntry;
